Better Sound object creation and BackgroundMusic implementation.

+++ src/com/Game/Callbacks/UpdateFunc.java
@@ -4,5 +4,5 @@ import com.Game.Scenes.Scene;
 
 @FunctionalInterface
 public interface UpdateFunc {
-    public void update(double dt , Scene currScene);
+    public void update(float dt , Scene currScene);
 }
